migrate auth diary and nutrition reference data to database

+++ backend/goodfood/src/main/kotlin/com/comp2850/goodfood/nutrition/DailyNutritionTrend.kt
@@ -4,7 +4,7 @@ import java.time.LocalDate
 
 data class DailyNutritionTrend(
     val date: LocalDate,
-    val totalCalories: Int,
+    val totalCalories: Double,
     val totalProtein: Double,
     val totalSugar: Double
 )
