+++ backend/goodfood/src/main/kotlin/com/comp2850/goodfood/nutrition/FoodNutrition.kt
@@ -2,7 +2,7 @@ package com.comp2850.goodfood.nutrition
 
 data class FoodNutrition(
     val foodName: String,
-    val calories: Int,
+    val calories: Double,
     val protein: Double,
     val sugar: Double
 )
