+++ backend/goodfood/src/main/kotlin/com/comp2850/goodfood/nutrition/NutritionSummary.kt
@@ -1,7 +1,7 @@
 package com.comp2850.goodfood.nutrition
 
 data class NutritionSummary(
-    val totalCalories: Int,
+    val totalCalories: Double,
     val totalProtein: Double,
     val totalSugar: Double,
     val matchedEntries: Int,
